+++ packages/devtools_app/lib/src/shared/feature_flags.dart
@@ -35,8 +35,8 @@ void setEnableExperiments() {
 @visibleForTesting
 bool get enableBeta => enableExperiments || !isExternalBuild;
 
-const _kMemoryOfflineExperiment =
-    bool.fromEnvironment('memory_offline_experiment', defaultValue: true);
+const _kMemoryDisconnectExperience =
+    bool.fromEnvironment('memory_disconnect_experience', defaultValue: true);
 
 // It is ok to have enum-like static only classes.
 // ignore: avoid_classes_with_only_static_members
@@ -56,12 +56,13 @@ abstract class FeatureFlags {
   /// https://github.com/flutter/devtools/issues/4564.
   static bool widgetRebuildStats = true;
 
-  /// Flag to enable offline data on memory screen.
+  /// Flag to enable viewing offline data on the memory screen when an app
+  /// disconnects.
   ///
   /// https://github.com/flutter/devtools/issues/5606
-  static const memoryOffline = _kMemoryOfflineExperiment;
+  static const memoryDisconnectExperience = _kMemoryDisconnectExperience;
 
-  /// Flag to enable save/load.
+  /// Flag to enable save/load for the Memory screen.
   ///
   /// https://github.com/flutter/devtools/issues/8019
   static bool memorySaveLoad = enableExperiments;
@@ -104,7 +105,7 @@ abstract class FeatureFlags {
   /// well.
   static final _allFlags = <String, bool>{
     'widgetRebuildStats': widgetRebuildStats,
-    'memoryOffline': memoryOffline,
+    'memoryOffline': memoryDisconnectExperience,
     'dapDebugging': dapDebugging,
     'loggingV2': loggingV2,
     'deepLinkIosCheck': deepLinkIosCheck,