+++ packages/devtools_app/lib/src/standalone_ui/standalone_screen.dart
@@ -4,6 +4,8 @@
 
 import 'package:flutter/material.dart';
 
+import '../shared/common_widgets.dart';
+import '../shared/globals.dart';
 import 'api/impl/dart_tooling_api.dart';
 import 'vs_code/flutter_panel.dart';
 
@@ -14,12 +16,24 @@ import 'vs_code/flutter_panel.dart';
 /// meaning that this screen will not be part of DevTools' normal navigation.
 /// The only way to access a standalone screen is directly from the url.
 enum StandaloneScreenType {
-  vsCodeFlutterPanel;
+  editorSidebar,
+  vsCodeFlutterPanel; // Legacy postMessage version.
 
   Widget get screen {
     return switch (this) {
       StandaloneScreenType.vsCodeFlutterPanel =>
         VsCodePostMessageSidebarPanel(PostMessageToolApiImpl.postMessage()),
+      StandaloneScreenType.editorSidebar => ValueListenableBuilder(
+          // TODO(dantup): Add a timeout here so if dtdManager.connection
+          //  doesn't complete after some period we can give some kind of
+          //  useful message.
+          valueListenable: dtdManager.connection,
+          builder: (context, data, _) {
+            return data == null
+                ? const CenteredCircularProgressIndicator()
+                : DtdEditorSidebarPanel(data);
+          },
+        ),
     };
   }
 }