+++ packages/devtools_app/test/inspector_v2/inspector_tree_test.dart
@@ -68,7 +68,6 @@ void main() {
         ),
       ),
     );
-    await tester.pumpAndSettle();
   }
 
   group('InspectorTreeController', () {