+++ packages/devtools_extensions/CHANGELOG.md
@@ -1,4 +1,10 @@
-## 0.2.2-wip
+## 0.3.0-dev.0
+* Add `ExtensionManager.copyToClipboard` method.
+* Add `DevToolsExtensionEventType.copyToClipboard` enum type.
+* Add `CopyToClipboardExtensionEvent` extension event type. 
+* Bump `web` dependency to `^1.0.0`.
+
+## 0.2.2
 * Load the IDE Theme from the extension URL instead of creating
 a placeholder `IdeTheme` object.